pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel Jerdhy Lapuz",email = "gabrieljerdhy.lapuz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "polars (>=1.32.1,<2.0.0)",
    "dlt (>=1.15.0) ; python_version >= \"3.10\" and python_version < \"3.15\"",
    "clickhouse-connect (>=0.8.18) ; python_version >= \"3.10\" and python_version < \"4.0\"",
    "certifi (>=2025.8.3,<2026.0.0)",
    "duckdb (>=1.3.2,<2.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "trackingmore-sdk-python (>=0.1.4) ; python_version >= \"3.10\" and python_version < \"4.0\""
]

[tool.poetry]
packages = [{include = "src", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
